Move test to fog of war tests
